requirements.txt
appdirs==1.4.4
asgiref==3.5.2
boto3==1.26.22
botocore==1.29.22
Django==4.1.3
django-bakery==0.13.1
django-tinymce==3.5.0
fs==2.4.16
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
